Upgrade monica
The officially published version of Monica is 4.x which is deprecated at
this point.  There is no Docker image for the "beta" version of Monica
(aka "Chandler" or 5.x), which means I needed to build and publish my
own copy from the Monica docker repository [1].

The command to generate the image was:

```
docker build -t slifty/monica:latest --platform=linux/amd64 .
docker push slifty/monica
```

Run inside of the `/5/fpm` directory of the docker repository.

Note there IS NO UPGRADE PATH so this needs to be run on a completely
fresh environment.

[1] https://github.com/monicahq/docker